+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## develop
 
+## v3.0.1
+
+Released Wednesday, 30th October 2019.
+
+### Fix
+
+* `OverlayEnv.IsExporter()` added, for interface compatibility
+
 ## v3.0.0
 
 Released Wednesday, 30th October 2019.